remove execution_helpers auxiliary file
